Update gititbk. com
In preview I seen a mistake but not sure what it means yet. I've been trying to study this but I really don't want the person to take it from me again. I'm studying really hard to learn all this and when I really started back a little over a year ago I had a better understanding of what I was doing. Thanks for any good feed back but I'm still studying and will try to update it today.